Make basic separation of javascript code on modules
